Moves view code to TypedTestView and keeps controller code in TypedTestController
